Fix creation date and DeltaT handling
DeltaTs are now relative to the smallest DeltaT (usually, the first plane in the first well).

The creation date is no longer used for DeltaT calculation, but will be stored
in the PlateAcquisition.StartTime (if it can be parsed). There will still be a logged
warning when the timezone is missing, but this will not impact DeltaT at all.