Remove unnecessary Python script for bumping Maven versions
The original intent of the script was to simplify the bump of the POM
versions as well as the static STABLE_VERSION field in the UpgradeChecker
class on release. In the context of the IDR fork of Bio-Formats,
additional functionality was added to manage versions for a custom groupId.

As the UpgradeChecker.STABLE_VERSION has been deprecated and the IDR fork
is no longer active, this script is redundant and can be fully replaced by
`mvn versions:set` which is the current documented workflow for Bio-Formats
releases.