ND2: add TextInfoItems to original metadata before parsing
This shouldn't affect parsing of individual key/value pairs,
but makes it easier to get each TextInfoItem as it is stored in the file
(and as it is presented in Elements). For TextInfoItems that are just a value
with no key (e.g. the timestamp in #3991), this makes sure that the value
is preserved when a valid key/value pair cannot be parsed.

Fixes #3991.