Metamorph TIFF: throw more explicit exceptions if parsing invalid TIFF
Fixes #3186.

If a TIFF without a valid IFD, or without the expected XML metadata,
is encountered, this throws FormatException with the name of the
affected file. That should be clearer than a NullPointerException with
no indication of which file is causing the problem.

In this particular reader, each TIFF needs to have readable metadata
in order to determine which specific plane (channel, Z section, etc.)
the file refers to. Per-file metadata is also used to calculate the
C, Z, and T sizes. That makes simply skipping problematic files
tricky, as it can throw off the dimensions for the whole dataset.